Created composition pattern branch with basic outline
