+++ model/LibraryElement.java
@@ -1,0 +1,7 @@
+package model;
+/**
+ * Component of the Composition pattern.
+  */
+public interface LibraryElement {
+    
+}
